wip combine radio and checkbox methods
