+++ services/ui-src/src/components/fields/CheckboxField.tsx
@@ -1,11 +1,10 @@
-import { useEffect, useState } from "react";
 import { useFormContext } from "react-hook-form";
 // components
 import { Box } from "@chakra-ui/react";
 import { ChoiceListField } from "components";
 // utils
 import { makeMediaQueryClasses } from "utils";
-import { AnyObject, InputChangeEvent, FieldChoice } from "types";
+import { AnyObject, FieldChoice } from "types";
 
 export const CheckboxField = ({
   name,
@@ -15,34 +14,9 @@ export const CheckboxField = ({
   ...props
 }: Props) => {
   const mqClasses = makeMediaQueryClasses();
-  const [fieldValues, setFieldValues] = useState<string[] | null>(
-    props.hydrate || null
-  );
 
   // get the form context
   const form = useFormContext();
-
-  // update local state
-  const onChangeHandler = (event: InputChangeEvent) => {
-    const checked = event.target.checked;
-    const clickedChoice = event.target.value;
-    const currentFieldValues = fieldValues || [];
-    const newFieldValues = checked
-      ? [...currentFieldValues, clickedChoice]
-      : currentFieldValues.filter((value) => value !== clickedChoice);
-    setFieldValues(newFieldValues);
-  };
-
-  useEffect(() => {
-    // update form data
-    if (fieldValues) {
-      form.setValue(name, fieldValues, { shouldValidate: true });
-      // update choice checked status
-      choices.forEach((choice: FieldChoice) => {
-        choice.checked = fieldValues.includes(choice.value);
-      });
-    }
-  }, [fieldValues]);
 
   const errorMessage = form?.formState?.errors?.[name]?.message;
 
@@ -53,7 +27,6 @@ export const CheckboxField = ({
         type={"checkbox"}
         label={label}
         choices={choices}
-        onChangeHandler={onChangeHandler}
         errorMessage={errorMessage}
         {...props}
       />
