+++ services/ui-src/src/components/fields/RadioField.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useState } from "react";
 import { useFormContext } from "react-hook-form";
 
 // components
@@ -6,7 +5,7 @@ import { Box } from "@chakra-ui/react";
 import { ChoiceListField } from "components";
 // utils
 import { makeMediaQueryClasses } from "utils";
-import { InputChangeEvent, AnyObject, FieldChoice } from "types";
+import { AnyObject, FieldChoice } from "types";
 
 export const RadioField = ({
   name,
@@ -17,22 +16,8 @@ export const RadioField = ({
 }: Props) => {
   const mqClasses = makeMediaQueryClasses();
 
-  const [fieldValue, setFieldValue] = useState<string[] | null>(null);
-
-  useEffect(() => {
-    if (fieldValue) {
-      form.setValue(name, fieldValue, { shouldValidate: true });
-    }
-  }, [fieldValue]);
-
   // get the form context
   const form = useFormContext();
-
-  // update form data
-  const onChangeHandler = (event: InputChangeEvent) => {
-    const optionValue = event.target.value;
-    setFieldValue([optionValue]);
-  };
 
   const errorMessage = form?.formState?.errors?.[name]?.message;
 
@@ -43,7 +28,6 @@ export const RadioField = ({
         type={"radio"}
         label={label}
         choices={choices}
-        onChangeHandler={onChangeHandler}
         errorMessage={errorMessage}
         {...props}
       />
