+++ services/ui-src/src/forms/mcpar/arp/arp.json
@@ -99,6 +99,66 @@
             }
           ]
         }
+      },
+      {
+        "id": "test2",
+        "type": "radio",
+        "props": {
+          "label": "Top level radio with nested fields",
+          "hint": "Pick the first option",
+          "choices": [
+            {
+              "name": "test2-o1",
+              "type": "choice",
+              "label": "Option 1, choice with nested child",
+              "value": "option1",
+              "checked": false,
+              "children": [
+                {
+                  "id": "test2-o1-c",
+                  "type": "radio",
+                  "props": {
+                    "label": "Nested child radio",
+                    "hint": "Pick the first option",
+                    "choices": [
+                      {
+                        "name": "test2-o1-c-o1",
+                        "type": "choice",
+                        "label": "Option 1, choice with nested child",
+                        "value": "option1-1",
+                        "checked": false,
+                        "children": [
+                          {
+                            "id": "test2-o1-c-o1-c",
+                            "type": "text",
+                            "props": {
+                              "label": "Nested child text field",
+                              "hint": "Enter some text or whatever"
+                            }
+                          }
+                        ]
+                      },
+                      {
+                        "name": "test2-o1-c-o2",
+                        "type": "choice",
+                        "label": "Option 2, no children",
+                        "value": "option2",
+                        "checked": false
+                      }
+                    ]
+                  }
+                }
+              ]
+            },
+            {
+              "name": "test2-o2",
+              "type": "choice",
+              "label": "Option 2, no children",
+              "value": "option2",
+              "checked": false
+            }
+          ]
+        }
       }
     ]
   }
