+++ services/ui-src/src/forms/mcpar/arp/arp.schema.ts
@@ -19,4 +19,19 @@ export default object({
     is: (value: any) => value && value.indexOf("option1-1") != -1,
     then: (schema: any) => schema.required("Field is required"),
   }),
+  // radio example
+  test2: array()
+    .required("Field is required")
+    .min(1, "Must make at least one selection"),
+  "test2-o1-c": array().when("test2", {
+    is: (value: any) => value && value.indexOf("option1") != -1,
+    then: (schema: any) =>
+      schema
+        .required("Field is required")
+        .min(1, "Must make at least one selection"),
+  }),
+  "test2-o1-c-o1-c": string().when("test2-o1-c", {
+    is: (value: any) => value && value.indexOf("option1-1") != -1,
+    then: (schema: any) => schema.required("Field is required"),
+  }),
 });
